+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 0.1.11
+- GitHub #5 : [iOS] Application crash when NFC tag is empty.
+- [iOS] Adding event when user cancels NFC reading session.
+- Minor changes.
+
+### 0.1.10
+- GitHub #1 : [ANDROID] Add support nfc tag id (serial number).
+
 ### 0.1.9
 - GitHub #1 : [ANDROID] Add support nfc tag id (serial number).